requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-Session==0.6.0
gpxpy==1.5.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
python-dotenv==1.0.1
gunicorn==21.2.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Werkzeug==3.0.1
requests==2.31.0 